tech: extract header from EntityDetailsContent
